pyproject.toml
[project]
name = "logPilot"
version = "0.4.0"
description = "A simple Kubernetes log viewer web app"
requires-python = ">=3.12"
dependencies = ["flask", "kubernetes", "requests"]
authors = [
  { name = "Jesse Goodier", email = "31039225+jessegoodier@users.noreply.github.com" },
]

[project.optional-dependencies]
dev = [
  "playwright",
  "gunicorn",
  "pytest",
  "pytest-playwright",
  "ruff",
  "isort",
  "packaging",
  "bump-my-version",
  "ruamel-yaml",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 119
target-version = "py312"

# ==== isort ====

[tool.isort]
profile = "black"                                 # Keep black profile for isort compatibility
line_length = 119
known_first_party = ["tests", "scripts", "hooks"]

# ==== bump-my-version ====

[tool.bumpversion]

commit = false
tag = false
message = "Bump version: {current_version} → {new_version}"

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = "version = \"{current_version}\""
replace = "version = \"{new_version}\""

[[tool.bumpversion.files]]
filename = "src/__init__.py"
search = "__version__ = \"{current_version}\""
replace = "__version__ = \"{new_version}\""

[[tool.bumpversion.files]]
filename = "src/main.py"
search = "__version__ = \"{current_version}\""
replace = "__version__ = \"{new_version}\""

[[tool.bumpversion.files]]
filename = "k8s/deployment.yaml"
search = "value: \"{current_version}\""
replace = "value: \"{new_version}\""

[[tool.bumpversion.files]]
filename = "charts/logpilot/Chart.yaml"
search = "version: {current_version}"
replace = "version: {new_version}"

[[tool.bumpversion.files]]
filename = "charts/logpilot/Chart.yaml"
search = "appVersion: {current_version}"
replace = "appVersion: {new_version}"

[[tool.bumpversion.files]]
filename = "charts/logpilot/src/__init__.py"
search = "__version__ = \"{current_version}\""
replace = "__version__ = \"{new_version}\""

[[tool.bumpversion.files]]
filename = "charts/logpilot/pyproject.toml"
search = "version = \"{current_version}\""
replace = "version = \"{new_version}\""